% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{DataPrep}
\alias{DataPrep}
\title{An ROEV Function}
\usage{
DataPrep(x)
}
\arguments{
\item{x}{data frame where the first four columns are time-mean-sd-n. The column names are not important, but the order of columns is.}
}
\description{
Prepare matrix with columns for Generation, Time step, Number of samples, Mean, Stdev, Ln Mean, Ln Stdev. Output from this function is ready to input into \code{TriPanelBC()}
}
\examples{
# DataPrep()
}
\keyword{DataPrep}
