% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawNorm.R
\name{DrawNorm}
\alias{DrawNorm}
\title{An ROEV Function}
\usage{
DrawNorm(mean, sd, base, hgt)
}
\arguments{
\item{mean}{Mean}

\item{sd}{Standard deviation}

\item{base}{Base}

\item{hgt}{Height}
}
\description{
DrawNorm, histogram of step rates
}
\examples{
# DrawNorm()
}
\keyword{normal}
