% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormScoreAsin.R
\name{NormScoreAsin}
\alias{NormScoreAsin}
\title{An ROEV Function}
\usage{
NormScoreAsin(rangemin, rangemax, scoremean, scoresd)
}
\arguments{
\item{rangemin}{range minimum}

\item{rangemax}{range maximum}

\item{scoremean}{score mean}

\item{scoresd}{score standard deviation}
}
\description{
Normalize score mean and standard deviation for a score range
}
\examples{
# NormScoreAsin()
}
\keyword{arcsin}
\keyword{deviation,}
\keyword{mean,}
\keyword{range,}
\keyword{standard}
