% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PalPanelBC.R
\name{PalPanelBC}
\alias{PalPanelBC}
\title{An ROEV Function}
\usage{
PalPanelBC(idrx, dr, xac, yac, bootn, mode, psize, equation, vadd)
}
\arguments{
\item{idrx}{matrix with -Inf removed, n, diff or rate, xac, yac}

\item{dr}{something}

\item{xac}{something}

\item{yac}{something}

\item{bootn}{something}

\item{mode}{something}

\item{psize}{something}

\item{equation}{something}

\item{vadd}{something}
}
\description{
PalPanel B or C, drawing and analyzing LDI or LRI temporal scaling
}
\examples{
# PalPanelBC()
}
\keyword{LDI}
\keyword{LRI}
