% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pc2Ar.R
\name{Pc2Ar}
\alias{Pc2Ar}
\title{An ROEV Function}
\usage{
Pc2Ar(percent)
}
\arguments{
\item{percent}{something}
}
\description{
Conversion from a percent to an arithmetic scale
}
\examples{
# Pc2Ar()
}
\keyword{percent}
