% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcsd2Arsd.R
\name{Pcsd2Arsd}
\alias{Pcsd2Arsd}
\title{An ROEV Function}
\usage{
Pcsd2Arsd(pc.m, pc.sd)
}
\arguments{
\item{pc.m}{something}

\item{pc.sd}{something}
}
\description{
Conversion of a standard deviation from a percent to an arithmetic scale
}
\examples{
# Pcsd2Arsd()
}
\keyword{pc.m,pc.sd}
