% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolSD.R
\name{PoolSD}
\alias{PoolSD}
\title{An ROEV Function}
\usage{
PoolSD(n1, n2, sd1, sd2)
}
\arguments{
\item{n1}{something}

\item{n2}{something}

\item{sd1}{something}

\item{sd2}{something}
}
\description{
Pool standard deviations weighted by n
}
\examples{
# PoolSD()
}
\keyword{deviation}
\keyword{standard}
