% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roev-package.R
\docType{package}
\name{roev-package}
\alias{roev}
\alias{roev-package}
\title{roev: Rates of Evolution}
\description{
Functions written by Philip D. Gingerich to accompany his book 'Rates of Evolution: A Quantitative Synthesis' \url{https://doi.org/10.1017/9781316711644} Originally deposited by at Gingerich https://doi.org/10.5061/dryad.1tn7123 and adapted for use by Ben Marwick.
}
\author{
\strong{Maintainer}: Ben Marwick \email{benmarwick@gmail.com} (\href{https://orcid.org/0000-0001-7879-4531}{ORCID})

Authors:
\itemize{
  \item Philip Gingerich (\href{https://orcid.org/0000-0002-1550-2674}{ORCID})
}

}
\keyword{internal}
