% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriPanelBC.R
\name{TriPanelBC}
\alias{TriPanelBC}
\title{An ROEV Function}
\usage{
TriPanelBC(idrx, dr, xac, yac, bootn, mode, psize, equation)
}
\arguments{
\item{idrx}{matrix with -Inf removed, n, diff or rate, xac, yac}

\item{dr}{something}

\item{xac}{something}

\item{yac}{something}

\item{bootn}{something}

\item{mode}{something}

\item{psize}{something}

\item{equation}{something}
}
\description{
TriPanel B or C, drawing and analyzing LDI or LRI temporal scaling
}
\examples{
# TriPanelBC()
}
\keyword{LDI}
\keyword{LRI}
