% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtRobLinBoot.R
\name{WgtRobLinBoot}
\alias{WgtRobLinBoot}
\title{An ROEV Function}
\usage{
WgtRobLinBoot(datmat, bootn)
}
\arguments{
\item{datmat}{something}

\item{bootn}{something}
}
\description{
Bootstrap weighted robust line fit
}
\examples{
# WgtRobLinBoot()
}
\keyword{WgtRobLinBoot}
