% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtRobLinFit.R
\name{WgtRobLinFit}
\alias{WgtRobLinFit}
\title{An ROEV Function}
\usage{
WgtRobLinFit(datmat)
}
\arguments{
\item{datmat}{matrix of line nos., weights, indep.var, and dep.var}
}
\description{
Weighted-robust-linear-model line fit
}
\examples{
# WgtRobLinFit()
}
\keyword{WgtRobLinFit}
