% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriPanelBC.R
\name{TriPanelBC}
\alias{TriPanelBC}
\title{An ROEV Function}
\usage{
TriPanelBC(idrx, dr = "r", xac, yac, bootn = 1000, mode, psize = 2, equation)
}
\arguments{
\item{idrx}{matrix input data}

\item{dr}{character; mode, "d" for difference (LDI) or "r" for rate (LRI), default is "r"}

\item{xac}{integer; panel placement coordinate x}

\item{yac}{integer; panel placement coordinate y}

\item{bootn}{integer; umber of bootstrap replications, default is 1000}

\item{mode}{character; as in "medians","all","mixed"}

\item{psize}{integer; circle size for points, default is 2}

\item{equation}{character; position as "normal","lower","none"}
}
\description{
TriPanel B or C, drawing and analyzing LDI or LRI temporal scaling
}
\examples{
# TriPanelBC()
}
\keyword{LDI}
\keyword{LRI}
